test(e2e): replace server-mode create-button check with existing-games-only check
The appshell_server_mode CI job asserted the "Save to server" create
button/caption that just got removed from /open for PUBLIC_HAS_SERVER=true.
Swaps it for a check that no create-game UI is present and the
play.questviva.com/open link is, matching the new behaviour.
